# backend/routes/spec/api-docs.yaml
openapi: "3.0.0"
info:
  version: 0.0.1
  title: Digital Realty Website Backend API
  description: |
    Provides the backend/bridge for the digital realty website
  contact:
    name: Propane Studios
servers:
  - url: /api/v1
    description: Version 1 of this server
security:
  - basicAuth: []
paths:
...
  /company:
    get:
      description: Get a list of all companies. A logged-in company must be in the role of a solutions architect, sales, or admin to receive the company list. Any other role gets denied
      summary: List all companies
      operationId: ListCompanies
      tags:
        - Company
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/Limit"
        - name: page
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/Page"
      responses:
        '200':
          description: A paged array of companies
          content:
            appliacation/json:
              schema:
                type: object
                properties:
                  company:
                    type: array
                    items:
                      $ref: "#/components/schemas/Company"
                  total:
                    $ref: "#/components/schemas/Total"
                  numPages:
                    $ref: "#/components/schemas/NumPages"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAuthorized"
        '500':
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      description: Create a company only with sales, solutions architect roles
      operationId: CreateCompany
      summary: Create a company
      tags:
        - Company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostCompany"
      responses:
       '201':
        description: company created
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SuccessCompany"
       '400':
        description: invalid request
        content:
          application/json:
            schema:  
                $ref: "#/components/schemas/InvalidRequest"
       '404':
          description: not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFound"
       '500':
        description: unexpected error
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Error"
  /company/my: 
    get:
      description: Get a list of all companies. A logged-in company must be in the role of a solutions architect, sales, customer, solutions engineer or admin to receive the company list. Any other role gets denied
      summary: List of companies the user is a part of
      tags:
        - Company
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/Limit"
        - name: page
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/Page"
      responses:
        '200':
          description: A paged array of companies
          content:
            appliacation/json:
              schema:
                  $ref: "#/components/schemas/Company"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAuthorized"
        '500':
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /company/{id}:
    get:
      description: Get a list of all companies where the role of sales, solutions architect
      summary: Get a company
      tags:
        - Company
      parameters:
        - name: id
          in: path
          description: The Company ID
          required: true
          schema:
            type: string
          example: 8aba095e-eb0d-410f-b1c1-371215fac0c3
      responses:
        '200':
          description: A list of companies that the logged in user has access to
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAuthorized"
        '404':
          description: The requested company, not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFound"
        '500':
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      description: update a company data
      operationId: update a company
      summary: Update company data
      tags:
        - Company
      parameters:
        - name: id
          in: path
          description: The company id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  description: The company name
                  type: string
      responses:
       '200':
          description: updates the data of the company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PutCompany"
       '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAuthorized"
       '404':
          description: not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFound"
       '500':
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /company/{id}/activate:
    put:
      description: Activate a company account
      summary: to activate the account it will need role as a sales and solutions architect
      parameters:
        - name: id
          in: path
          description: The company id
          required: true
          schema:
            type: string
      tags:
        - Company
      responses:
       '200':
        description: updates the data from archived
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PutCompanyActivate"
  /company/{id}/:
    delete:
      description: Delete a company. The logged in company must be in role solutions architect, sales, customer, solutions engineer or admin to delete (or archive) a company. Any other role gets denied
      summary: Delete a company alows only for sales and solutions architect roles (archive)
      parameters:
        - name: id
          in: path
          description: To delete company it will need UUID
          required: true
          schema:
            $ref: '#/components/schemas/IdFields'
      tags:
        - Company
      responses:
        "200":
          description: Successfully deleted (archived) the user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessCompany"
        '500':
          description: An error occurred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error" 
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
    Total:
      type: number
      description: total number of records
    NumPages:
      type: number
      description: total number of pages given the limit
    Limit:
      type: number
      description: Total number of records per page (max 100)
    Page:
      description: Which page to start on (relative to limit)
      type: integer
      format: int32
...
    CompanyNameField:
      type: string
      description: name of the company
...
    Company:
      required:
        - name
      properties:
        id:
          $ref: '#/components/schemas/IdFields'
        name:
          $ref: '#/components/schemas/CompanyNameField'
        createdBy:
          $ref: '#/components/schemas/CreatedBy'
...
    SuccessCompany:
      required:
        - company
      properties:
        company:
          type: string
          description: Contains the success message
...
    InvalidRequest:
      required:
        - error
      properties:
        error:
          type: string
          description: Contains a message detailing why the request was invalid
    Error:
      required:
        - error
      properties:
        error:
          type: string
          description: contains a message about the failure
...
    PutCompany:
      properties:
        id:
          $ref: '#/components/schemas/IdFields'
        name:
          $ref: "#/components/schemas/CompanyNameField"
    PutCompanyActivate:
      properties:
        id:
          $ref: '#/components/schemas/IdFields'
    PostCompany:
      required:
        - name
      properties:
        name:
          $ref: "#/components/schemas/CompanyNameField"
...
    CreatedBy:
      type: string
      description: id for creating user
      example: b0d0ab14-4069-48df-9258-86647a0d7d95
    IdFields:
      type: string
      description: Object id (UUID)
      example: b0d0ab14-4069-48df-9258-86647a0d7d95
...
    NotFound:
      required:
        - error
      properties:
        error:
          type: string
          example: "Not Found"
    NotAuthorized:
      required:
        - error
      properties:
        error:
          type: string
          example: "Not Authorized"